+++ crypto_sign/dilithium5/clean/poly.c
@@ -548,14 +548,14 @@ void PQCLEAN_DILITHIUM5_CLEAN_polyeta_pack(uint8_t *r, const poly *a) {
     DBENCH_START();
 
     for (i = 0; i < N / 8; ++i) {
-        t[0] = ETA - a->coeffs[8 * i + 0];
-        t[1] = ETA - a->coeffs[8 * i + 1];
-        t[2] = ETA - a->coeffs[8 * i + 2];
-        t[3] = ETA - a->coeffs[8 * i + 3];
-        t[4] = ETA - a->coeffs[8 * i + 4];
-        t[5] = ETA - a->coeffs[8 * i + 5];
-        t[6] = ETA - a->coeffs[8 * i + 6];
-        t[7] = ETA - a->coeffs[8 * i + 7];
+        t[0] = (uint8_t) (ETA - a->coeffs[8 * i + 0]);
+        t[1] = (uint8_t) (ETA - a->coeffs[8 * i + 1]);
+        t[2] = (uint8_t) (ETA - a->coeffs[8 * i + 2]);
+        t[3] = (uint8_t) (ETA - a->coeffs[8 * i + 3]);
+        t[4] = (uint8_t) (ETA - a->coeffs[8 * i + 4]);
+        t[5] = (uint8_t) (ETA - a->coeffs[8 * i + 5]);
+        t[6] = (uint8_t) (ETA - a->coeffs[8 * i + 6]);
+        t[7] = (uint8_t) (ETA - a->coeffs[8 * i + 7]);
 
         r[3 * i + 0]  = (t[0] >> 0) | (t[1] << 3) | (t[2] << 6);
         r[3 * i + 1]  = (t[2] >> 2) | (t[3] << 1) | (t[4] << 4) | (t[5] << 7);
@@ -615,11 +615,11 @@ void PQCLEAN_DILITHIUM5_CLEAN_polyt1_pack(uint8_t *r, const poly *a) {
     DBENCH_START();
 
     for (i = 0; i < N / 4; ++i) {
-        r[5 * i + 0] = (a->coeffs[4 * i + 0] >> 0);
-        r[5 * i + 1] = (a->coeffs[4 * i + 0] >> 8) | (a->coeffs[4 * i + 1] << 2);
-        r[5 * i + 2] = (a->coeffs[4 * i + 1] >> 6) | (a->coeffs[4 * i + 2] << 4);
-        r[5 * i + 3] = (a->coeffs[4 * i + 2] >> 4) | (a->coeffs[4 * i + 3] << 6);
-        r[5 * i + 4] = (a->coeffs[4 * i + 3] >> 2);
+        r[5 * i + 0] = (uint8_t) (a->coeffs[4 * i + 0] >> 0);
+        r[5 * i + 1] = (uint8_t) ((a->coeffs[4 * i + 0] >> 8) | (a->coeffs[4 * i + 1] << 2));
+        r[5 * i + 2] = (uint8_t) ((a->coeffs[4 * i + 1] >> 6) | (a->coeffs[4 * i + 2] << 4));
+        r[5 * i + 3] = (uint8_t) ((a->coeffs[4 * i + 2] >> 4) | (a->coeffs[4 * i + 3] << 6));
+        r[5 * i + 4] = (uint8_t) (a->coeffs[4 * i + 3] >> 2);
     }
 
     DBENCH_STOP(*tpack);
@@ -672,26 +672,26 @@ void PQCLEAN_DILITHIUM5_CLEAN_polyt0_pack(uint8_t *r, const poly *a) {
         t[6] = (1 << (D - 1)) - a->coeffs[8 * i + 6];
         t[7] = (1 << (D - 1)) - a->coeffs[8 * i + 7];
 
-        r[13 * i + 0]  =  t[0];
-        r[13 * i + 1]  =  t[0] >>  8;
-        r[13 * i + 1] |=  t[1] <<  5;
-        r[13 * i + 2]  =  t[1] >>  3;
-        r[13 * i + 3]  =  t[1] >> 11;
-        r[13 * i + 3] |=  t[2] <<  2;
-        r[13 * i + 4]  =  t[2] >>  6;
-        r[13 * i + 4] |=  t[3] <<  7;
-        r[13 * i + 5]  =  t[3] >>  1;
-        r[13 * i + 6]  =  t[3] >>  9;
-        r[13 * i + 6] |=  t[4] <<  4;
-        r[13 * i + 7]  =  t[4] >>  4;
-        r[13 * i + 8]  =  t[4] >> 12;
-        r[13 * i + 8] |=  t[5] <<  1;
-        r[13 * i + 9]  =  t[5] >>  7;
-        r[13 * i + 9] |=  t[6] <<  6;
-        r[13 * i + 10]  =  t[6] >>  2;
-        r[13 * i + 11]  =  t[6] >> 10;
-        r[13 * i + 11] |=  t[7] <<  3;
-        r[13 * i + 12]  =  t[7] >>  5;
+        r[13 * i + 0]  =  (uint8_t) t[0];
+        r[13 * i + 1]  =  (uint8_t) (t[0] >>  8);
+        r[13 * i + 1] |=  (uint8_t) (t[1] <<  5);
+        r[13 * i + 2]  =  (uint8_t) (t[1] >>  3);
+        r[13 * i + 3]  =  (uint8_t) (t[1] >> 11);
+        r[13 * i + 3] |=  (uint8_t) (t[2] <<  2);
+        r[13 * i + 4]  =  (uint8_t) (t[2] >>  6);
+        r[13 * i + 4] |=  (uint8_t) (t[3] <<  7);
+        r[13 * i + 5]  =  (uint8_t) (t[3] >>  1);
+        r[13 * i + 6]  =  (uint8_t) (t[3] >>  9);
+        r[13 * i + 6] |=  (uint8_t) (t[4] <<  4);
+        r[13 * i + 7]  =  (uint8_t) (t[4] >>  4);
+        r[13 * i + 8]  =  (uint8_t) (t[4] >> 12);
+        r[13 * i + 8] |=  (uint8_t) (t[5] <<  1);
+        r[13 * i + 9]  =  (uint8_t) (t[5] >>  7);
+        r[13 * i + 9] |=  (uint8_t) (t[6] <<  6);
+        r[13 * i + 10]  =  (uint8_t) (t[6] >>  2);
+        r[13 * i + 11]  =  (uint8_t) (t[6] >> 10);
+        r[13 * i + 11] |=  (uint8_t) (t[7] <<  3);
+        r[13 * i + 12]  =  (uint8_t) (t[7] >>  5);
     }
 
     DBENCH_STOP(*tpack);
